Update dependencies
Ref. None

+++ rflx/spark.py
@@ -4,18 +4,26 @@ import typing as ty
 
 from pydantic import BaseModel
 
+# TODO(eng/recordflux/RecordFlux#1359): Replace ty.* by collections.abc.*
+# Sequence and Mapping are imported from collections.abc as importing them
+# from typing is deprecated. However pydantic does not support the imported
+# version from collections.abc. To fix that typing is imported as ty and the
+# typing versions of Sequence and Mapping are used in classes that derive
+# from pydantic.BaseModel.
+# This is only relevant for Python 3.8.
 
-class ProofAttempt(BaseModel):
+
+class ProofAttempt(BaseModel):  # type: ignore[misc]
     result: str
     steps: int
-    time: int
+    time: float
 
 
-class CheckTreeElement(BaseModel):
+class CheckTreeElement(BaseModel):  # type: ignore[misc]
     proof_attempts: ty.Mapping[str, ProofAttempt]
 
 
-class Proof(BaseModel):
+class Proof(BaseModel):  # type: ignore[misc]
     file: str
     line: int
     col: int
@@ -23,5 +31,5 @@ class Proof(BaseModel):
     check_tree: ty.Sequence[CheckTreeElement]
 
 
-class SPARKFile(BaseModel):
+class SPARKFile(BaseModel):  # type: ignore[misc]
     proof: ty.Sequence[Proof]
